MuPiHat automation

+++ scripts/mupihat/mupihat_automation.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-SOUND_FILE="/home/dietpi/MuPiBox/sysmedia/sound/low.mp3"
+SOUND_FILE="/home/dietpi/MuPiBox/sysmedia/sound/low.wav"
 JSON_FILE="/tmp/mupihat.json"
 BATTERY_LOW="/home/dietpi/MuPiBox/sysmedia/images/battery_low.jpg"
 CONFIG="/etc/mupibox/mupiboxconfig.json"
@@ -17,16 +17,19 @@ if [ "${BAT_CONNECTED}" -eq 1 ]; then
 	while true; do
 		if [ -f ${JSON_FILE} ]; then
 			IBUS=$(jq -r '.IBus' ${JSON_FILE})
-			if [ "$IBUS" -eq 0 ]; then
-				STATE=$(jq -r '.Bat_Stat' ${JSON_FILE})
-				if [ "${STATE}" = "LOW" ]; then
-					play_sound
-					echo "Battery state low"
-				elif [ "${STATE}" = "SHUTDOWN" ]; then
-					/usr/local/bin/mupibox/mupi_shutdown.sh ${BATTERY_LOW}
-					echo "Battery state to low - shutdown initiated"
-					service mupi_powerled stop 
-					poweroff
+			VBUS=$(jq -r '.VBus' ${JSON_FILE})
+            if [ "$VBUS" -le 1000 ]; then
+				if [ "$IBUS" -eq 0 ]; then
+					STATE=$(jq -r '.Bat_Stat' ${JSON_FILE})
+					if [ "${STATE}" = "LOW" ]; then
+						play_sound
+						echo "Battery state low"
+					elif [ "${STATE}" = "SHUTDOWN" ]; then
+						/usr/local/bin/mupibox/mupi_shutdown.sh ${BATTERY_LOW}
+						echo "Battery state to low - shutdown initiated"
+						service mupi_powerled stop 
+						poweroff
+					fi
 				fi
 			fi
 		fi
